fixed: refactor
